chore: split App.vue into multiple components
